+++ solidity/contracts/converter/interfaces/IBancorConverterUpgrader.sol
@@ -5,5 +5,5 @@ import './IBancorConverter.sol';
     Bancor Converter Upgrader interface
 */
 contract IBancorConverterUpgrader {
-    function upgrade(IBancorConverter _converter, bytes32 _version) public;
+    function upgrade(bytes32 _version) public;
 }